replaced old customer with new customer
