+++ lib/index.js
@@ -10,8 +10,6 @@ require('./common')(app);
 
 require('./models')(app);
 
-require('./models2')(app);
-
 require('./routes')(app);
 
 require('./server')(app);
