--- lib/models/customer/sanitize.js
@@ -1,13 +1,0 @@
-module.exports = function(schema) {
-
-  schema.path('name').trim();
-
-  schema.path('email').trim().lowercase();
-
-  schema.path('company').trim();
-
-  schema.path('address.street').trim();
-
-  schema.path('address.city').trim();
-
-};
