--- lib/models/customer/virtual.js
@@ -1,18 +1,0 @@
-var gravatar = require('gravatar');
-
-module.exports = function(schema) {
-
-  schema.virtual('image').get(function() {
-    return {
-      url: generateUrl(this.email)
-    };
-  });
-
-};
-
-function generateUrl(email) {
-  return gravatar.url(email || '', {
-    default: 'mm',
-    size: '500'
-  });
-}
