+++ lib/models/index.js
@@ -1,6 +1,12 @@
+var monk     = require('monk');
+var lodash   = require('lodash');
 var mongoose = require('mongoose');
 
+var Customer = require('./customer');
+
 module.exports = function(app) {
+
+  app.context.db = monk(app.storage.url);
 
   mongoose.connect(app.storage.url);
 
@@ -16,5 +22,14 @@ module.exports = function(app) {
 
   require('./order')(app);
 
+  lodash.merge(app.context, {
+    Customer: bind(Customer, app.context.db.get('customers'))
+  });
+
 };
 
+function bind(model, collection) {
+  model.collection = collection;
+
+  return model;
+}
